refactor: centralize site product content
- Move feature, capability, pricing, provider, screenshot, FAQ, and alternatives arrays into lib/content.ts
- Derive JSON-LD featureList from FEATURES summaries to prevent inventory drift
