+++ site/components/gallery.tsx
@@ -1,40 +1,4 @@
-const screenshots = [
-  {
-    src: "/screenshots/review.example.webp",
-    alt: "PR Agent automated AI code review on a GitHub pull request using /review",
-    caption: "/review",
-    width: 806,
-    height: 480,
-  },
-  {
-    src: "/screenshots/describe.example.webp",
-    alt: "PR Agent AI-generated pull request description using /describe",
-    caption: "/describe",
-    width: 806,
-    height: 643,
-  },
-  {
-    src: "/screenshots/ask.example.webp",
-    alt: "PR Agent answering a code question on a GitHub pull request with /ask",
-    caption: "/ask",
-    width: 806,
-    height: 745,
-  },
-  {
-    src: "/screenshots/review-security.example.webp",
-    alt: "PR Agent security code review summary on a GitHub pull request with /review-security",
-    caption: "/review-security",
-    width: 806,
-    height: 639,
-  },
-  {
-    src: "/screenshots/review-quality.example.webp",
-    alt: "PR Agent code quality review on a GitHub pull request with /review-quality",
-    caption: "/review-quality",
-    width: 536,
-    height: 726,
-  },
-];
+import { SCREENSHOTS } from "@/lib/content";
 
 export function Gallery() {
   return (
@@ -54,7 +18,7 @@ export function Gallery() {
         </p>
 
         <div className="space-y-6">
-          {screenshots.map((shot) => (
+          {SCREENSHOTS.map((shot) => (
             <figure key={shot.caption}>
               <div className="rounded-md border border-neutral-200 overflow-hidden">
                 <img
